deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
scikit-learn==1.2.2
tensorflow==2.12.0
keras
numpy
Pillow
transformers